app/views: Moved navbar from App.js
The NavBar is separated from the App.js into a separate file to keep the
code readable
